% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flip_card.R
\name{card}
\alias{card}
\title{Generate a Single Flip Card}
\usage{
card(
  title,
  subheading = NULL,
  description = NULL,
  img = NULL,
  flip_axis = "Y",
  front = NULL,
  back = NULL,
  img_alt = NULL,
  id = "flip-card",
  sass_options = sass::sass_options(output_style = "compressed"),
  ...
)
}
\arguments{
\item{title}{character card title}

\item{subheading}{character card subheading}

\item{description}{character card description}

\item{img}{character filepath}

\item{flip_axis}{character flip on X or Y plane, defaults to Y}

\item{front}{taglist for more complex sides}

\item{back}{taglist for more complex sides}

\item{img_alt}{character}

\item{id}{character unique id for card or grid of cards. This
allows for muliple configurations on a single page}

\item{...}{additional options passed to sass}
}
\value{

}
\description{
Generate a Single Flip Card
}
