% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_cards.R
\name{cards_grid}
\alias{cards_grid}
\title{Card Grid}
\usage{
cards_grid(
  df,
  width,
  height,
  id = "grid-cards",
  sass_options = sass::sass_options(output_style = "compressed"),
  ...
)
}
\arguments{
\item{df}{}

\item{id}{character unique id for card or grid of cards. This
allows for muliple configurations on a single page}

\item{sass_options}{}

\item{...}{}
}
\value{

}
\description{
Card Grid
}
