% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module.R
\name{gallery}
\alias{gallery}
\title{Gallery Module Server}
\usage{
gallery(
  input,
  output,
  session,
  images,
  gallery = c("lightbox", "photoswipe"),
  ...
)
}
\arguments{
\item{input}{}

\item{output}{}

\item{session}{}

\item{images}{}

\item{gallery}{}

\item{...}{}
}
\value{

}
\description{
Gallery Module Server
}
