% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exif.R
\name{get_exif}
\alias{get_exif}
\title{Image Exif Extraction}
\usage{
get_exif(file, exiftool = NULL, opt = "-common")
}
\arguments{
\item{file}{charater string including path to file}

\item{exiftool}{default location of exiftool}

\item{opt}{character options sent to exiftool}
}
\value{
data.frame
}
\description{
Image Exif Extraction
}
