% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lightbox.R
\name{lightbox_gallery}
\alias{lightbox_gallery}
\title{Create Lightbox Gallery}
\usage{
lightbox_gallery(df, gallery, css, display = "block")
}
\arguments{
\item{df}{data.frame containing}

\item{gallery}{character label identifying gallery}

\item{css}{character file path to css file. If missing default style is used}

\item{display}{character}
}
\description{
Create Lightbox Gallery
}
